+++ tests/mixins/test_WeightColumnMixin.py
@@ -1,9 +1,175 @@
+from copy import deepcopy
+
 import narwhals as nw
 import numpy as np
 import pytest
 
 from tests.test_data import create_df_2
+from tests.utils import assert_frame_equal_dispatch, dataframe_init_dispatch
 from tubular.mixins import WeightColumnMixin
+
+
+class TestCreateDummyWeightsColumn:
+    @pytest.mark.parametrize("library", ["pandas", "polars"])
+    @pytest.mark.parametrize("i", [-1, 0, 1, 2, 3, 4])
+    def test_new_column_output(
+        self,
+        library,
+        i,
+    ):
+        """Test dummy weights column created as expected"""
+
+        obj = WeightColumnMixin()
+
+        df_dict = {
+            "a": [1, 2, 3, 4],
+        }
+
+        for j in range(-2, i):
+            if j == -2:
+                pass
+            elif j == -1:
+                df_dict["dummy_weights_column"] = [1, 2, 1, 1]
+            else:
+                df_dict[f"dummy_weights_column_{j}"] = [1, 2, 1, 1]
+
+        df = dataframe_init_dispatch(dataframe_dict=df_dict, library=library)
+
+        expected_dict = deepcopy(df_dict)
+
+        expected_new_col = (
+            "dummy_weights_column" if i == -1 else f"dummy_weights_column_{i}"
+        )
+
+        expected_dict[expected_new_col] = [1, 1, 1, 1]
+
+        expected = dataframe_init_dispatch(
+            dataframe_dict=expected_dict,
+            library=library,
+        )
+
+        output = obj._create_dummy_weights_column(df, backend=library)
+
+        assert_frame_equal_dispatch(expected, output)
+
+    @pytest.mark.parametrize("library", ["pandas", "polars"])
+    @pytest.mark.parametrize("i", [0, 1, 2, 3, 4])
+    def test_warnings_raised_for_existing_dummy_weight_columns(
+        self,
+        library,
+        i,
+    ):
+        """Test warnings are raised when dummy weights columns already exist"""
+
+        obj = WeightColumnMixin()
+
+        df_dict = {
+            "a": [1, 2, 3, 4],
+        }
+
+        for j in range(-2, i):
+            if j == -2:
+                pass
+            elif j == -1:
+                df_dict["dummy_weights_column"] = [1, 2, 1, 1]
+            else:
+                df_dict[f"dummy_weights_column_{j}"] = [1, 2, 1, 1]
+
+        df = dataframe_init_dispatch(dataframe_dict=df_dict, library=library)
+
+        with pytest.warns(
+            UserWarning,
+        ) as record:
+            obj._create_dummy_weights_column(df, backend=library)
+
+        previous_dummy_weights_columns = [
+            "dummy_weights_column",
+            *[f"dummy_weights_column_{i}" for i in range(i)],
+        ]
+        final_dummy_weights_columns = [
+            f"dummy_weights_column_{i}" for i in range(i + 1)
+        ]
+
+        assert len(record) == len(final_dummy_weights_columns)
+
+        for j, msg in enumerate(
+            [
+                f"{previous_dummy_weights_column} already exists in X but is not all 1, attempting to add {final_dummy_weights_column}"
+                for previous_dummy_weights_column, final_dummy_weights_column in zip(
+                    previous_dummy_weights_columns,
+                    final_dummy_weights_columns,
+                )
+            ],
+        ):
+            assert msg in str(record[j].message)
+
+    @pytest.mark.parametrize("library", ["pandas", "polars"])
+    @pytest.mark.parametrize("i", [0, 1, 2, 3, 4])
+    def test_existing_column_used_if_possible(
+        self,
+        library,
+        i,
+    ):
+        """Test dummy weights column created as expected"""
+
+        obj = WeightColumnMixin()
+
+        df_dict = {
+            "a": [1, 2, 3, 4],
+        }
+
+        for j in range(-1, i):
+            if j == -1:
+                df_dict["dummy_weights_column"] = (
+                    [1, 1, 2, 1] if i != j + 1 else [1, 1, 1, 1]
+                )
+            else:
+                df_dict[f"dummy_weights_column_{j}"] = (
+                    [1, 1, 1, 1] if i != j + 1 else [1, 1, 1, 1]
+                )
+
+        df = dataframe_init_dispatch(dataframe_dict=df_dict, library=library)
+
+        expected_dict = deepcopy(df_dict)
+
+        expected = dataframe_init_dispatch(
+            dataframe_dict=expected_dict,
+            library=library,
+        )
+
+        output = obj._create_dummy_weights_column(df, backend=library)
+
+        assert_frame_equal_dispatch(expected, output)
+
+    @pytest.mark.parametrize("library", ["pandas", "polars"])
+    def test_errors_if_too_many_failed_attempts(
+        self,
+        library,
+    ):
+        """Test that error is raised if suitable column name not found quickly"""
+
+        obj = WeightColumnMixin()
+
+        df_dict = {
+            "a": [1, 2, 3, 4],
+        }
+
+        for j in range(-2, 5):
+            if j == -2:
+                pass
+            elif j == -1:
+                df_dict["dummy_weights_column"] = [1, 2, 1, 1]
+            else:
+                df_dict[f"dummy_weights_column_{j}"] = [1, 2, 1, 1]
+
+        df = dataframe_init_dispatch(dataframe_dict=df_dict, library=library)
+
+        msg = "Taking too long to find unused name for dummy weights column, consider renaming columns like 'dummy_weights_column' in X"
+        with pytest.raises(
+            RuntimeError,
+            match=msg,
+        ):
+            obj._create_dummy_weights_column(df, backend=library)
 
 
 class TestCheckAndSetWeight: