+++ tubular/capping.py
@@ -190,30 +190,25 @@ class BaseCappingTransformer(BaseNumericTransformer, WeightColumnMixin):
             Required for pipeline.
 
         """
-        if self.weights_column:
-            WeightColumnMixin.check_weights_column(self, X, self.weights_column)
 
         super().fit(X, y)
 
-        self.quantile_capping_values = {}
+        backend = nw.get_native_namespace(X)
 
-        native_backend = nw.get_native_namespace(X)
+        weights_column = self.weights_column
+        if self.weights_column is None:
+            X = WeightColumnMixin._create_dummy_weights_column(
+                X,
+                backend=backend.__name__,
+                return_native=False,
+            )
+            weights_column = "dummy_weights_column"
+        WeightColumnMixin.check_weights_column(self, X, weights_column)
+
+        self.quantile_capping_values = {}
 
         if self.quantiles is not None:
             for col in self.columns:
-                if self.weights_column is None:
-                    weights_column = "dummy_weights_column"
-                    X = X.with_columns(
-                        nw.new_series(
-                            name="dummy_weights_column",
-                            values=[1] * len(X),
-                            backend=native_backend,
-                        ),
-                    )
-
-                else:
-                    weights_column = self.weights_column
-
                 cap_values = self.prepare_quantiles(
                     X,
                     self.quantiles[col],
@@ -693,8 +688,21 @@ class OutOfRangeNullTransformer(BaseCappingTransformer):
         """
         super().fit(X=X, y=y)
 
-        if self.weights_column:
-            WeightColumnMixin.check_weights_column(self, X, self.weights_column)
+        backend = nw.get_native_namespace(X)
+
+        weights_column = self.weights_column
+        if self.weights_column is None:
+            X = WeightColumnMixin._create_dummy_weights_column(
+                X,
+                backend=backend.__name__,
+                return_native=False,
+            )
+            weights_column = "dummy_weights_column"
+        WeightColumnMixin.check_weights_column(self, X, weights_column)
+
+        # need to overwrite attr for fit method to work
+        original_weights_column = weights_column
+        self.weights_column = weights_column
 
         if self.quantiles:
             BaseCappingTransformer.fit(self, X=X, y=y)
@@ -708,5 +716,8 @@ class OutOfRangeNullTransformer(BaseCappingTransformer):
                 f"{self.classname()}: quantiles not set so no fitting done in OutOfRangeNullTransformer",
                 stacklevel=2,
             )
+
+        # restore attr
+        self.weights_column = original_weights_column
 
         return self